fix: Move pareseXLS to the Convert class so it can be shared
